HD-267. Исправление начисления мастерства пользователю.

+++ Hinode.Izumi.Services/GameServices/PropertyService/Records/MasteryXpPropertyRecord.cs
@@ -1,12 +1,12 @@
 ﻿namespace Hinode.Izumi.Services.GameServices.PropertyService.Records
 {
     public record MasteryXpPropertyRecord(
-        long Mastery0,
-        long Mastery50,
-        long Mastery100,
-        long Mastery150,
-        long Mastery200,
-        long Mastery250)
+        double Mastery0,
+        double Mastery50,
+        double Mastery100,
+        double Mastery150,
+        double Mastery200,
+        double Mastery250)
     {
         public MasteryXpPropertyRecord() : this(default, default, default, default, default, default)
         {
